pyproject.toml
[project]
name = "lendflow-takehome-assignment"
version = "0.1.0"
description = "My solution to the Lendflow backend take-home assignment"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.7",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "psycopg>=3.2.6",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "django-extensions>=3.2.3",
    "pytest>=8.3.5",
    "pytest-django>=4.10.0",
    "ruff>=0.11.2",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP", "SIM"]
ignore = ["F401"]
